refactor(dashboard): migrate SearchFilters to Nuxt UI components and remove map page
- Replace custom dropdown implementations with `USelect` components in `SearchFilters`
- Replace custom search input with `UInput` component
- Replace custom favorite button with `UButton` component
- Add `ALL_CATEGORIES_VALUE` constant for category filtering
- Simplify category selection logic with computed display label
- Remove manual dropdown state management (`showDistanceDropdown`, `showCategoryDropdown`)
- Remove
